+++ shadycompass/rules/http_busting.py
