+++ shadycompass/rules/imap_scanner.py
