+++ shadycompass/rules/pop_scanner.py
