+++ tests/rules/test_rules_dns_scanner.py
