+++ tests/rules/test_rules_ldap_scanner.py
