+++ tests/rules/test_rules_pop_scanner.py
