+++ tests/rules/test_rules_smb_scanner.py
